pyproject.toml
[project]
name = "torch_candle"
version = "0.1.0"
description = "A PyTorch-compatible API with Candle backend"
authors = [
    { name = "Antigravity", email = "antigravity@example.com" }
]
dependencies = [
    "candle-pyo3",
    "numpy",
    "pytest>=8.3.5",
]
requires-python = ">=3.8, <3.11"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
